refactor: API 계층 추상화
